BAEL-574 adding reporters as beans instead of config classes
