+++ spring-cloud/spring-cloud-bootstrap/gateway/src/main/java/com/baeldung/spring/cloud/bootstrap/gateway/GatewayApplication.java
@@ -1,6 +1,9 @@
 package com.baeldung.spring.cloud.bootstrap.gateway;
 
+import com.netflix.appinfo.InstanceInfo;
+import com.netflix.discovery.EurekaClient;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
@@ -8,8 +11,13 @@ import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 import org.springframework.cloud.netflix.ribbon.RibbonClientSpecification;
 import org.springframework.cloud.netflix.ribbon.SpringClientFactory;
 import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
+import org.springframework.cloud.sleuth.metric.SpanMetricReporter;
+import org.springframework.cloud.sleuth.zipkin.HttpZipkinSpanReporter;
+import org.springframework.cloud.sleuth.zipkin.ZipkinProperties;
+import org.springframework.cloud.sleuth.zipkin.ZipkinSpanReporter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.client.RestTemplate;
+import zipkin.Span;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,23 +26,50 @@ import java.util.List;
 @EnableZuulProxy
 @EnableEurekaClient
 public class GatewayApplication {
-        public static void main(String[] args) {
-                SpringApplication.run(GatewayApplication.class, args);
-        }
+    public static void main(String[] args) {
+        SpringApplication.run(GatewayApplication.class, args);
+    }
 
-        @Autowired(required = false)
-        private List<RibbonClientSpecification> configurations = new ArrayList<>();
+    @Autowired(required = false)
+    private List<RibbonClientSpecification> configurations = new ArrayList<>();
+    @Autowired
+    private EurekaClient eurekaClient;
+    @Autowired
+    private SpanMetricReporter spanMetricReporter;
+    @Autowired
+    private ZipkinProperties zipkinProperties;
+    @Value("${spring.sleuth.web.skipPattern}")
+    private String skipPattern;
 
-        @Bean
-        @LoadBalanced
-        RestTemplate restTemplate() {
-                return new RestTemplate();
-        }
+    @Bean
+    @LoadBalanced
+    RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
 
-        @Bean
-        public SpringClientFactory springClientFactory() {
-                SpringClientFactory factory = new SpringClientFactory();
-                factory.setConfigurations(this.configurations);
-                return factory;
-        }
+    @Bean
+    public SpringClientFactory springClientFactory() {
+        SpringClientFactory factory = new SpringClientFactory();
+        factory.setConfigurations(this.configurations);
+        return factory;
+    }
+
+    @Bean
+    public ZipkinSpanReporter makeZipkinSpanReporter() {
+        return new ZipkinSpanReporter() {
+            private HttpZipkinSpanReporter delegate;
+            private String baseUrl;
+
+            @Override
+            public void report(Span span) {
+                InstanceInfo instance = eurekaClient.getNextServerFromEureka("zipkin", false);
+                if (!(baseUrl != null && instance.getHomePageUrl().equals(baseUrl))) {
+                    baseUrl = instance.getHomePageUrl();
+                    delegate = new HttpZipkinSpanReporter(baseUrl, zipkinProperties.getFlushInterval(), zipkinProperties.getCompression().isEnabled(), spanMetricReporter);
+                    if (!span.name.matches(skipPattern)) delegate.report(span);
+                }
+                if (!span.name.matches(skipPattern)) delegate.report(span);
+            }
+        };
+    }
 }
