+++ spring-cloud/spring-cloud-bootstrap/svc-book/src/main/resources/bootstrap.properties
@@ -8,3 +8,4 @@ eureka.client.serviceUrl.defaultZone=http://discUser:discPassword@localhost:8082
 
 spring.sleuth.sampler.percentage=1.0
 spring.sleuth.web.skipPattern=(^cleanup.*)
+spring.zipkin.locator.discovery.enabled=true
